Changed what buttons would be displayed on pages
Now there is a button on only the storybook page with the relevant game. Button is not present on any page except storybook page 21.

The milestones in the Milestone checklist now link to each relevant game if there is one. The current games are linked to milestones 10, 20, and 22. If you click the milestone a description of the milestone will come up, there is an option to rate the child's performance, or you can click the button to play the game version. If there is not a game version of the milestone, then no 'Play the Game' button is present.